+++ actions/list.py
@@ -245,13 +245,14 @@ class ListPreviewAction(html5.Span):
 	def __init__(self, module, handler, actionName, *args, **kwargs ):
 		super(ListPreviewAction, self ).__init__(*args, **kwargs)
 
+		self.urls = conf["defaultPreview"]
+
 		self.urlCb = html5.Select()
 		self.appendChild(self.urlCb)
 
 		btn = html5.ext.Button(translate("Preview"), callback=self.onClick)
 		btn["class"] = "icon preview"
 		self.appendChild(btn)
-		self.urls = None
 
 		self["disabled"] = True
 		self.isDisabled = True
@@ -268,7 +269,7 @@ class ListPreviewAction(html5.Span):
 			self.urls = {x: x for x in self.urls}
 
 		if not isinstance(self.urls, dict) or len(self.urls.keys()) == 1:
-			self.urlCb["style"]["display"] = "none"
+			self.urlCb.hide()
 			return
 
 		for name, url in self.urls.items():
@@ -277,7 +278,7 @@ class ListPreviewAction(html5.Span):
 			o.appendChild(html5.TextNode(name))
 			self.urlCb.appendChild(o)
 
-		self.urlCb["style"]["display"] = ""
+		self.urlCb.show()
 
 	def onAttach(self):
 		super(ListPreviewAction,self).onAttach()
@@ -287,9 +288,12 @@ class ListPreviewAction(html5.Span):
 		if module in conf["modules"].keys():
 			moduleConfig = conf["modules"][module]
 
-			self.urls = moduleConfig.get("preview", moduleConfig.get("previewurls"))
-			if self.urls:
-				self.rebuildCB()
+			urls = moduleConfig.get("preview", moduleConfig.get("previewurls"))
+			if urls:
+				self.urls = urls
+
+		if self.urls:
+			self.rebuildCB()
 
 	def onDetach(self):
 		self.parent().parent().selectionChangedEvent.unregister(self)
@@ -307,7 +311,7 @@ class ListPreviewAction(html5.Span):
 				self.isDisabled = True
 
 	def onClick(self, sender=None):
-		if self.urls is None:
+		if not self.urls:
 			return
 
 		selection = self.parent().parent().getCurrentSelection()
@@ -336,35 +340,35 @@ class ListPreviewAction(html5.Span):
 			eval("""window.open('""" + newUrl + """', '""" + target + """');""")
 
 	@staticmethod
-	def isSuitableFor( module, handler, actionName ):
-		if module is None or module not in conf["modules"].keys():
+	def isSuitableFor(module, handler, actionName):
+		if module is None or module not in conf["modules"]:
 			return False
+
+		modConf = conf["modules"][module]
 
 		correctAction = actionName=="preview"
 		correctHandler = handler == "list" or handler.startswith("list.")
 		hasAccess = conf["currentUser"] and ("root" in conf["currentUser"]["access"] or module+"-view" in conf["currentUser"]["access"])
-		isDisabled = module is not None and "disabledFunctions" in conf["modules"][module].keys() and conf["modules"][module]["disabledFunctions"] and "view" in conf["modules"][module]["disabledFunctions"]
-		isAvailable = conf["modules"][module].get("preview", conf["modules"][module].get("previewurls"))
+		isDisabled = "disabledFunctions" in modConf and modConf["disabledFunctions"] and "view" in conf["modules"][module]["disabledFunctions"]
+		isAvailable = bool(modConf.get("preview", modConf.get("previewurls", conf["defaultPreview"])))
 
 		return correctAction and correctHandler and hasAccess and not isDisabled and isAvailable
 
-actionDelegateSelector.insert( 2, ListPreviewAction.isSuitableFor, ListPreviewAction )
+actionDelegateSelector.insert(2, ListPreviewAction.isSuitableFor, ListPreviewAction)
 
 
-class ListPreviewInlineAction( html5.ext.Button ):
-	def __init__(self, *args, **kwargs ):
-		super( ListPreviewInlineAction, self ).__init__( translate("Preview"), *args, **kwargs )
-		self["class"] = "icon preview"
-		self["disabled"] = True
-		self.urls = None
+class ListPreviewInlineAction(html5.Span):
+	def __init__(self, *args, **kwargs):
+		super(ListPreviewInlineAction, self ).__init__()
+		self.hide()
 
 	def onAttach(self):
-		super( ListPreviewInlineAction,self ).onAttach()
+		super(ListPreviewInlineAction, self).onAttach()
 		self.parent().parent().selectionChangedEvent.register( self )
 
 	def onDetach(self):
 		self.parent().parent().selectionChangedEvent.unregister( self )
-		super( ListPreviewInlineAction, self ).onDetach()
+		super(ListPreviewInlineAction, self).onDetach()
 
 	def onSelectionChanged(self, table, selection):
 		if self.parent().parent().selectMode:
@@ -398,10 +402,11 @@ class ListPreviewInlineAction( html5.ext.Button ):
 		correctHandler = handler == "list" or handler.startswith("list.")
 		hasAccess = conf["currentUser"] and ("root" in conf["currentUser"]["access"] or module+"-view" in conf["currentUser"]["access"])
 		isDisabled = module is not None and "disabledFunctions" in conf["modules"][module].keys() and conf["modules"][module]["disabledFunctions"] and "view" in conf["modules"][module]["disabledFunctions"]
+		generallyDisabled = not bool(conf["modules"][module].get("disableInternalPreview", not conf["internalPreview"]))
 
-		return correctAction and correctHandler and hasAccess and not isDisabled
+		return correctAction and correctHandler and hasAccess and not isDisabled and generallyDisabled
 
-actionDelegateSelector.insert( 1, ListPreviewInlineAction.isSuitableFor, ListPreviewInlineAction )
+actionDelegateSelector.insert(1, ListPreviewInlineAction.isSuitableFor, ListPreviewInlineAction)
 
 
 class CloseAction( html5.ext.Button ):
@@ -590,7 +595,6 @@ class ListSelectFilterAction( html5.ext.Button ):
 	def __init__(self, *args, **kwargs ):
 		super( ListSelectFilterAction, self ).__init__( translate("Select Filter"), *args, **kwargs )
 		self["class"] = "icon selectfilter"
-		self.urls = None
 		self.filterSelector = None
 
 	def onAttach(self):